seperating out the email and suggestions to a spreadheet controller
