+++ web/router.ex
@@ -20,7 +20,7 @@ defmodule App.Router do
     get "/", HomepageController, :index
     post "/", HomepageController, :show
     get "/filter", HomepageController, :query
-    post "/email", HomepageController, :submit_email
+    post "/email", SpreadsheetController, :submit_email
     post "/like/:article_id", HomepageController, :like
     post "/dislike/:article_id", HomepageController, :dislike
     get "/article/:id", ArticleController, :show
